fix: fix linting errors
