+++ convex/quizSessions.ts
@@ -1,10 +1,9 @@
 import { v } from 'convex/values';
 
 import { internal } from './_generated/api';
-import { Id } from './_generated/dataModel';
 import { query } from './_generated/server';
-import { getCurrentUserOrThrow } from './users';
 import { mutation } from './triggers';
+import { getCurrentUserOrThrow } from './users';
 
 /**
  * Quiz Sessions
